Final version of guessing game.
With this version now we check the user's guess in a loop and direct
the user to guess a number closer to random number. If the user's
guess is smaller than the random number we give a feedback to user to
guess higher number. If the user's guess is bigger than the random
number we give a feedback to user to guess lower number. If the user
guesses correct we simply say "Correct!" and break the loop with changing
the boolean variable found.